+++ src/app/layout.tsx
@@ -1,7 +1,9 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import MainWrapper from './MainWrapper';
-import Provider from '@/components/layout/Provider';
+import MainWrapper from '../components/layout/MainWrapper';
+import Provider from '@/app/Provider';
+import Header from '@/components/layout/Header';
+import Footer from '@/components/layout/Footer';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -19,7 +21,9 @@ export default function RootLayout({
     <html lang='en'>
       <body className={inter.className}>
         <Provider>
+          <Header />
           <MainWrapper>{children}</MainWrapper>
+          <Footer />
         </Provider>
       </body>
     </html>
