--- src/app/main/page.tsx
@@ -1,16 +1,0 @@
-import Navbar from '@/components/layout/Navbar';
-import React from 'react';
-import Crawling from '@/components/crawling';
-import Footer from '@/components/layout/Footer';
-
-const Main = () => {
-  return (
-    <div>
-      <Navbar />
-      <Crawling />
-      <Footer />
-    </div>
-  );
-};
-
-export default Main;
