+++ src/components/layout/Header.tsx
@@ -1,9 +1,8 @@
-import React from 'react'
+import React from 'react';
+import Navbar from './Navbar';
 
 const Header = () => {
-  return (
-    <div>Header</div>
-  )
-}
+  return <Navbar />;
+};
 
-export default Header
+export default Header;
